pyproject.toml
[tool.poetry]
name = "lunch-log"
version = "1.0.0"
description = "Office Lunch Receipt Management and Recommendation System Backend"
authors = ["Mohammed Rishad <mohammedrishad.vk@gmail.com>"]
readme = "README.md"
packages = [{include = "lunch_log"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.13"
djangorestframework = "3.15.1"
drf-spectacular = "0.27.2"
psycopg2-binary = "2.9.9"
django-cors-headers = "4.3.1"
ruff = "0.4.9"
django-storages = "1.14.3"
boto3 = "1.34.127"
pillow = "10.3.0"
django-filter = "24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
